requirements.txt
torch>=1.10.0
torchvision>=0.11.1
Pillow>=8.0.0
matplotlib>=3.4.0
tqdm>=4.62.0
